# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Job Portal API Documentation",
        "description": "API for managing Job Offers",
        "contact": {
            "email": "support@musicbox.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://127.0.0.1:8000",
            "description": "Job Portal API Server"
        }
    ],
    "paths": {
...
        "/api/jobs": {
            "get": {
                "tags": [
                    "Offres d'emploi"
                ],
                "summary": "Liste de toutes les offres d'emploi",
                "description": "Récupère la liste complète de toutes les offres d'emploi disponibles.",
                "operationId": "6b307032ccc7685af33c78dee9152c22",
                "responses": {
                    "200": {
                        "description": "Liste des offres d'emploi récupérée avec succès.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "properties": {
                                            "id": {
                                                "type": "integer",
                                                "example": 1
                                            },
                                            "title": {
                                                "type": "string",
                                                "example": "Développeur Full Stack (Laravel/Vue)"
                                            },
                                            "description": {
                                                "type": "string",
                                                "example": "Création et maintenance de nos applications web."
                                            },
                                            "salary": {
                                                "type": "number",
                                                "format": "float",
                                                "example": 55000
                                            },
                                            "location": {
                                                "type": "string",
                                                "example": "Paris, France"
                                            },
                                            "status": {
                                                "type": "string",
                                                "enum": [
                                                    "Ouvert",
                                                    "Fermé",
                                                    "En attente"
                                                ],
                                                "example": "Ouvert"
                                            },
                                            "created_at": {
                                                "type": "string",
                                                "format": "date-time"
                                            },
                                            "updated_at": {
                                                "type": "string",
                                                "format": "date-time"
                                            }
                                        },
                                        "type": "object"
                                    }
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Offres d'emploi"
                ],
                "summary": "Créer une nouvelle offre d'emploi",
                "description": "Enregistre une nouvelle offre d'emploi. L'accès est limité aux rôles 'admin' et 'emp'.",
                "operationId": "7ec8a93a554353d753e8c5e3a610e3c8",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "title",
                                    "description",
                                    "entreprise",
                                    "ville",
                                    "user_id"
                                ],
                                "properties": {
                                    "title": {
                                        "type": "string",
                                        "example": "Développeur Back-end Senior"
                                    },
                                    "description": {
                                        "type": "string",
                                        "example": "Responsable de l'architecture API et des bases de données."
                                    },
                                    "entreprise": {
                                        "type": "string",
                                        "example": "Tech Innov S.A."
                                    },
                                    "ville": {
                                        "type": "string",
                                        "example": "Casablanca"
                                    },
                                    "user_id": {
                                        "description": "ID de l'utilisateur (employeur/admin) créant l'offre.",
                                        "type": "integer",
                                        "example": 2
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Offre d'emploi créée avec succès.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Job created"
                                        },
                                        "data": {
                                            "properties": {
                                                "title": {
                                                    "type": "string",
                                                    "example": "Développeur Back-end Senior"
                                                },
                                                "description": {
                                                    "type": "string",
                                                    "example": "Responsable de l'architecture API et des bases de données."
                                                },
                                                "entreprise": {
                                                    "type": "string",
                                                    "example": "Tech Innov S.A."
                                                },
                                                "ville": {
                                                    "type": "string",
                                                    "example": "Casablanca"
                                                },
                                                "user_id": {
                                                    "type": "integer",
                                                    "example": 2
                                                },
                                                "id": {
                                                    "type": "integer",
                                                    "example": 12
                                                },
                                                "created_at": {
                                                    "type": "string",
                                                    "format": "date-time"
                                                },
                                                "updated_at": {
                                                    "type": "string",
                                                    "format": "date-time"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Interdit (Non autorisé).",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "not exiscte"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Erreur de validation des données."
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
        "/api/jobs/search": {
            "get": {
                "tags": [
                    "Offres d'emploi"
                ],
                "summary": "Rechercher des offres d'emploi",
                "description": "Permet de rechercher des offres d'emploi en utilisant des critères optionnels (titre, entreprise, ville).",
                "operationId": "4256b2e2755d716a244e334806e03d60",
                "parameters": [
                    {
                        "name": "title",
                        "in": "query",
                        "description": "Recherche partielle par titre de l'offre d'emploi.",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "example": "Développeur"
                        }
                    },
                    {
                        "name": "entreprise",
                        "in": "query",
                        "description": "Recherche partielle par nom d'entreprise.",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "example": "Tech Innov"
                        }
                    },
                    {
                        "name": "ville",
                        "in": "query",
                        "description": "Recherche partielle par ville.",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "example": "Casablanca"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Liste des offres d'emploi correspondant aux critères de recherche.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "properties": {
                                            "id": {
                                                "type": "integer",
                                                "example": 1
                                            },
                                            "title": {
                                                "type": "string",
                                                "example": "Développeur Full Stack"
                                            },
                                            "description": {
                                                "type": "string",
                                                "example": "Description de l'offre."
                                            },
                                            "entreprise": {
                                                "type": "string",
                                                "example": "Tech Innov S.A."
                                            },
                                            "ville": {
                                                "type": "string",
                                                "example": "Casablanca"
                                            },
                                            "created_at": {
                                                "type": "string",
                                                "format": "date-time"
                                            },
                                            "updated_at": {
                                                "type": "string",
                                                "format": "date-time"
                                            }
                                        },
                                        "type": "object"
                                    }
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "name": "Bearer Token Authentication",
                "in": "header",
                "bearerFormat": "Sanctum Token",
                "scheme": "Bearer"
            }
        }
    },
...
}